% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{arrangePlots}
\alias{arrangePlots}
\title{Arange multiple PCA matrices into a plot grid}
\usage{
arrangePlots(pcs, row.cols, color.cols)
}
\arguments{
\item{pcs}{PCA matrices.}

\item{row.cols}{Variables to use for rows.}

\item{color.cols}{Variables to use for columns.}
}
\value{
A patchwork plot object
}
\description{
Arange multiple PCA matrices into a plot grid
}
