% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrection.R
\name{batchCorrection}
\alias{batchCorrection}
\title{Perform batch correction using the selected method.}
\usage{
batchCorrection(
  d,
  quantCol,
  batch1,
  batch2 = NULL,
  covariate.cols = NULL,
  rowsName = "precursor",
  columnsName = "replicate",
  bc.method = "limma",
  return.format = "long"
)
}
\arguments{
\item{d}{Long formated dataframe where each row is an observation for a single condition.}

\item{quantCol}{Name of the column with quantative values.}

\item{batch1}{Name of the column to use for the first batch.}

\item{batch2}{Name of the column to use for the second batch. (optional). Using a second batch is not supported with ComBat}

\item{covariate.cols}{The names of the covariate columns. (optional)}

\item{rowsName}{Name of the column to be row names in the wide matrix. Default is 'precursor'}

\item{columnsName}{Name of the column to be column names in the wide matrix. Default is 'replicate'}

\item{bc.method}{The batch correction column to use. One of c('combat', 'limma'). Default is 'limma'}

\item{return.format}{How should the resulting dataframe be formated?
'Long': Long formated dataframe. (default)
'Wide': Wide formated dataframe.
'Matrix': Wide fromated numeric matrix.}
}
\value{
Batch correced dataframe.
}
\description{
Perform batch correction using the selected method.
}
