% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCleaning.R
\name{longToMatrix}
\alias{longToMatrix}
\title{Convert long dataframe to wide numeric matrix.}
\usage{
longToMatrix(d, valuesFrom, rowsName, columnsName)
}
\arguments{
\item{d}{Long formatted dataframe}

\item{valuesFrom}{The name of the column with values that should be in the matrix.}

\item{rowsName}{Name of the column to be row names in the matrix}

\item{columnsName}{Name of the column to be column names in the matrix}
}
\value{
A numeric matrix.
}
\description{
Convert long dataframe to wide numeric matrix.
}
