% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCleaning.R
\name{parseAcquiredTime}
\alias{parseAcquiredTime}
\title{Convert a date string into POSIX format}
\usage{
parseAcquiredTime(d, acquiredTimeCol = "acquiredTime")
}
\arguments{
\item{acquiredTimeCol}{Name of column with with the UTC date to parse}

\item{dat}{Dataframe}
}
\value{
Dataframe with acquiredTimeCol as an R POSIXlt object.
}
\description{
Convert a date string into POSIX format
}
