% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcAnalysis.R
\name{pcAnalysis}
\alias{pcAnalysis}
\title{Perform PCA analysis on a column in a long formatted dataframe}
\usage{
pcAnalysis(
  dat,
  quantCol,
  rowsName = "precursor",
  columnsName = "replicate",
  x.pc = 1,
  y.pc = 2,
  scale = TRUE,
  na.rm = FALSE,
  return.all.pcs = FALSE
)
}
\arguments{
\item{dat}{Long formatted dataframe}

\item{quantCol}{The name of the column with quantative values to analyze}

\item{rowsName}{Name of the column to be row names in the wide matrix}

\item{columnsName}{Name of the column to be column names in the wide matrix}

\item{x.pc}{PC 1}

\item{y.pc}{PC 2}

\item{scale}{Scale argument passed to prcomp}

\item{na.rm}{Remove rows with NAs? Default is False.}

\item{return.all.pcs}{Return the full PC matrix instead of just 2 dimensions.}
}
\value{
list with 3 slots:
      pc: The pc dataframe
      x.lab: The x label for x.pc
      y.lab: The y label for y.pc
}
\description{
Perform PCA analysis on a column in a long formatted dataframe
}
