% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PCAScatterPlot}
\alias{PCAScatterPlot}
\title{Make PCA scatter plot.}
\usage{
PCAScatterPlot(
  pc,
  dat.metadata,
  color.col,
  plot.title = NULL,
  show.ylab = T,
  show.xlab = T,
  show.legend = F,
  legend.maxLabelPerCol = 12
)
}
\arguments{
\item{pc}{A PCA matrix.}

\item{dat.metadata}{A dataframe mapping replicate names to metadata terms to use in plots.}

\item{color.col}{Column to use for color aesthetic.}

\item{plot.title}{Plot title. By default there is no title.}

\item{show.ylab}{Show y axis label?}

\item{show.xlab}{Show x axis label?}

\item{show.legend}{Show legend for color aesthetic?}

\item{legend.maxLabelPerCol}{An integer indicating the maximum labels per column in the legend.}
}
\value{
A ggplot object.
}
\description{
Make PCA scatter plot.
}
