% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCleaning.R
\name{pivotLonger}
\alias{pivotLonger}
\title{Convert long dataframe to wide dataframe.}
\usage{
pivotLonger(d, valuesFrom, rowsName, columnsName)
}
\arguments{
\item{d}{Long formatted dataframe}

\item{valuesFrom}{The name of the column with values to be spread into columns.}

\item{rowsName}{Name of the column(s) to keep for each row.}

\item{columnsName}{Name of the column to be column names.}
}
\value{
A wide format dataframe.
}
\description{
Convert long dataframe to wide dataframe.
}
