% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCleaning.R
\name{readDataFile}
\alias{readDataFile}
\title{Parse tsv file and save it as a serilized R object.}
\usage{
readDataFile(file.basename, ext = ".tsv", sep = "\\t")
}
\arguments{
\item{file.basename}{The basename of the tsv/Rda file.}

\item{ext}{The file extension. Default is '.tsv'}

\item{sep}{The deliminator. Default is '\\t'}
}
\description{
If the serilized file already exists, it is read instead of the tsv file
}
