% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCleaning.R
\name{readDirectLFQMatrix}
\alias{readDirectLFQMatrix}
\title{Parse DirectLFQ matrix into a long formated dataframe.}
\usage{
readDirectLFQMatrix(fname)
}
\arguments{
\item{fname}{The DirectLFQ matrix.}
}
\description{
Also save intermediate .Rda file similar to rDIAUtils::readDataFile.
}
\seealso{
[rDIAUtils::readDataFile]
}
