% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCleaning.R
\name{readWideMetadata}
\alias{readWideMetadata}
\title{Read metadata and replicates from batch database.}
\usage{
readWideMetadata(conn)
}
\arguments{
\item{conn}{An initialized database connection}
}
\value{
Replicate and metadata dataframe
}
\description{
Metadata is converted from long format used in database and converted to wide format.
The values from the sampleMetadataTypes table are then used to convert each column
into the appropriate date type.
}
