% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smallestDifferentStrings}
\alias{smallestDifferentStrings}
\title{Remove the substring which is shared by the begining and end of all strings in a CharacterVector.}
\usage{
smallestDifferentStrings(cStrings, verbose = TRUE)
}
\arguments{
\item{cStrings}{The CharacterVector.}

\item{verbose}{Print verbose output?}
}
\value{
CharacterVector with common begining and ending removed.
}
\description{
Remove the substring which is shared by the begining and end of all strings in a CharacterVector.
}
