% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{arrangePlots}
\alias{arrangePlots}
\title{Arange multiple PCA matrices into a plot grid}
\usage{
arrangePlots(
  pcs,
  row.cols,
  color.cols,
  dat.metadata,
  x.pc = 1,
  y.pc = 2,
  interactive = F,
  legend.maxLabelPerCol = 12,
  nrow = length(color.cols),
  return.list = FALSE,
  ...
)
}
\arguments{
\item{pcs}{A list of PCA matrices.}

\item{row.cols}{Variables to use for rows.}

\item{color.cols}{Variables to use for columns.}

\item{dat.metadata}{A dataframe mapping replicate names to metadata terms to use in plots.}

\item{x.pc}{PC to plot on x axis}

\item{y.pc}{PC to plot on y axis}

\item{interactive}{Use ggiraph::geom_point_interactive?}

\item{legend.maxLabelPerCol}{An integer indicating the maximum labels per column in the legend.}

\item{nrow}{Number of rows in plot layout.
If there are more than 1 color.cols nrow must be equal to length(color.cols)}

\item{return.list}{Return a list of ggplot objects instead of a patchwork object.}

\item{...}{Additional arguments passed to plot theme()}
}
\value{
A patchwork plot object
}
\description{
Arange multiple PCA matrices into a plot grid
}
